moved sign-in form into header and made home and onboarding component
