+++ packages/frontend/src/components/Onboarding.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Onboarding = () => {
+  return (
+    <div>
+      <div>This is the onboarding page</div>
+      <div>Next button</div>
+    </div>
+  )
+}
+
+export default Onboarding
