+++ packages/frontend/tsconfig.json
@@ -10,7 +10,7 @@
     "target": "es5",
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": false,
   },
   "include": [
     "src/**/*"
